Feature/client side unwinding (#26)
Add support for client side unwinding

If client side unwinding is enabled, then when there is a native/crashpad crash, the call stack will be resolved by the device (client) before submission to the Backtrace server. This allows better resolution of call stacks
when symbols for native libraries are not available, e.g for syscalls or opaque libraries.